+++ src/main/java/ru/ai/sin/service/impl/CategoryService.java
@@ -10,15 +10,22 @@ import java.util.List;
 public interface CategoryService {
 
     // ---------- GET METHODS ----------
-    CategoryDTO getById(Long id);
+    CategoryDTO getById(
+            Long id);
 
-    List<CategoryDTO> getAll(long page, long size);
+    List<CategoryDTO> getAll(
+            int page, int size);
 
     // ---------- POST METHODS ----------
-    CategoryDTO create(AddCategoryReq addCategoryReq);
+    CategoryDTO create(
+            AddCategoryReq addCategoryReq);
 
-    CategoryDTO setNameById(Long id, SetCategoryNameReq setCategoryNameReq);
-    CategoryDTO setSkillsById(Long id, SetCategorySkillsReq setCategorySkillsReq);
+    CategoryDTO setNameById(
+            long id,
+            SetCategoryNameReq setCategoryNameReq);
+    CategoryDTO setSkillsById(
+            long id,
+            SetCategorySkillsReq setCategorySkillsReq);
 
     // ---------- POST METHODS ----------
     CategoryDTO deleteById(Long id);
